# database/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package database

import (
	"database/sql/driver"
	"fmt"

	"github.com/jackc/pgx/v5/pgtype"
)

type Userstatus string

const (
	UserstatusActive   Userstatus = "Active"
	UserstatusInactive Userstatus = "Inactive"
)

func (e *Userstatus) Scan(src interface{}) error {
	switch s := src.(type) {
	case []byte:
		*e = Userstatus(s)
	case string:
		*e = Userstatus(s)
	default:
		return fmt.Errorf("unsupported scan type for Userstatus: %T", src)
	}
	return nil
}

type NullUserstatus struct {
	Userstatus Userstatus
	Valid      bool // Valid is true if Userstatus is not NULL
}

// Scan implements the Scanner interface.
func (ns *NullUserstatus) Scan(value interface{}) error {
	if value == nil {
		ns.Userstatus, ns.Valid = "", false
		return nil
	}
	ns.Valid = true
	return ns.Userstatus.Scan(value)
}

// Value implements the driver Valuer interface.
func (ns NullUserstatus) Value() (driver.Value, error) {
	if !ns.Valid {
		return nil, nil
	}
	return string(ns.Userstatus), nil
}

type User struct {
	Userid     int32
	Firstname  string
	Lastname   string
	Email      string
	Phone      pgtype.Text
	Age        pgtype.Int4
	UserStatus NullUserstatus
}